Add english translations for everything.
This moves all visible japanese text out into a resource file, and adds
english translations for every string.

Also adds a new Setting called Language, which can be JP or EN, but
for now it defaults to JP and there's no way to change it from within
the plugin interface, so text will still always be japanese for now.